@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class AdminController < ApplicationController
   layout :determine_layout
   before_action :cache_setup
@@ -8,60 +10,61 @@ class AdminController < ApplicationController
   USERS_URL = "#{LinkedData::Client.settings.rest_url}/users"
   ONTOLOGY_URL = lambda { |acronym| "#{ADMIN_URL}ontologies/#{acronym}" }
   PARSE_LOG_URL = lambda { |acronym| "#{ONTOLOGY_URL.call(acronym)}/log" }
-  REPORT_NEVER_GENERATED = "NEVER GENERATED"
+  REPORT_NEVER_GENERATED = 'NEVER GENERATED'
 
   def index
     @users = LinkedData::Client::Models::User.all
     if session[:user].nil? || !session[:user].admin?
-      redirect_to :controller => 'login', :action => 'index', :redirect => '/admin'
+      redirect_to controller: 'login', action: 'index', redirect: '/admin'
     else
-      render action: "index"
+      render action: 'index'
     end
   end
 
   def update_info
-    response = {update_info: Hash.new, errors: '', success: '', notices: ''}
+    response = { update_info: {}, errors: '', success: '', notices: '' }
     json = LinkedData::Client::HTTP.get("#{ADMIN_URL}update_info", params, raw: true)
 
     begin
       update_info = JSON.parse(json)
 
-      if update_info["error"]
-        response[:errors] = update_info["error"]
+      if update_info['error']
+        response[:errors] = update_info['error']
+        response[:update_info] = update_info
       else
         response[:update_info] = update_info
-        response[:notices] = update_info["notes"] if update_info["notes"]
-        response[:success] = "Update info successfully retrieved"
+        response[:notices] = update_info['notes'] if update_info['notes']
+        response[:success] = 'Update info successfully retrieved'
       end
-    rescue Exception => e
+    rescue StandardError => e
       response[:errors] = "Problem retrieving update info - #{e.message}"
     end
-    render :json => response
+    render json: response
   end
 
   def update_check_enabled
     enabled = LinkedData::Client::HTTP.get("#{ADMIN_URL}update_check_enabled", {}, raw: false)
-    render :json => enabled
+    render json: enabled
   end
 
   def submissions
     @submissions = nil
-    @acronym = params["acronym"]
-    @ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params["acronym"]).first
+    @acronym = params['acronym']
+    @ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params['acronym']).first
     begin
       submissions = @ontology.explore.submissions
-      @submissions = submissions.sort {|a,b| b.submissionId <=> a.submissionId }
+      @submissions = submissions.sort { |a, b| b.submissionId <=> a.submissionId }
     rescue
       @submissions = []
     end
-    render :partial => "layouts/ontology_report_submissions"
+    render partial: 'layouts/ontology_report_submissions'
   end
 
   def parse_log
-    @acronym = params["acronym"]
-    @parse_log = LinkedData::Client::HTTP.get(PARSE_LOG_URL.call(params["acronym"]), {}, raw: false)
+    @acronym = params['acronym']
+    @parse_log = LinkedData::Client::HTTP.get(PARSE_LOG_URL.call(params['acronym']), {}, raw: false)
     ontologies_report = _ontologies_report
-    ontology = ontologies_report[:ontologies][params["acronym"].to_sym]
+    ontology = ontologies_report[:ontologies][params['acronym'].to_sym]
     @log_file_path = ''
 
     if ontology
@@ -69,96 +72,95 @@ class AdminController < ApplicationController
       @log_file_path = /#{params["acronym"]}\/\d+\/[-a-zA-Z0-9_]+\.log$/.match(full_log_file_path)
     else
       @parse_log = "No record exists for ontology #{params["acronym"]}"
-      @log_file_path = "None"
+      @log_file_path = 'None'
     end
-    render action: "parse_log"
+    render action: 'parse_log'
   end
 
   def clearcache
-    response = {errors: '', success: ''}
+    response = { errors: '', success: '' }
 
     if @cache.respond_to?(:flush_all)
       begin
         @cache.flush_all
-        response[:success] = "UI cache successfully flushed"
-      rescue Exception => e
+        response[:success] = 'UI cache successfully flushed'
+      rescue StandardError => e
         response[:errors] = "Problem flushing the UI cache - #{e.class}: #{e.message}"
       end
     else
       response[:errors] = "The UI cache does not respond to the 'flush_all' command"
     end
-    render :json => response
+    render json: response
   end
 
   def resetcache
-    response = {errors: '', success: ''}
+    response = { errors: '', success: '' }
 
     if @cache.respond_to?(:reset)
       begin
         @cache.reset
-        response[:success] = "UI cache connection successfully reset"
-      rescue Exception => e
+        response[:success] = 'UI cache connection successfully reset'
+      rescue StandardError => e
         response[:errors] = "Problem resetting the UI cache connection - #{e.message}"
       end
     else
       response[:errors] = "The UI cache does not respond to the 'reset' command"
     end
-    render :json => response
+    render json: response
   end
 
   def clear_goo_cache
-    response = {errors: '', success: ''}
+    response = { errors: '', success: '' }
 
     begin
-      response_raw = LinkedData::Client::HTTP.post("#{ADMIN_URL}clear_goo_cache", params, raw: true)
-      response[:success] = "Goo cache successfully flushed"
-    rescue Exception => e
+      LinkedData::Client::HTTP.post("#{ADMIN_URL}clear_goo_cache", params, raw: true)
+      response[:success] = 'Goo cache successfully flushed'
+    rescue StandardError => e
       response[:errors] = "Problem flushing the Goo cache - #{e.class}: #{e.message}"
     end
-    render :json => response
+    render json: response
   end
 
   def clear_http_cache
-    response = {errors: '', success: ''}
+    response = { errors: '', success: '' }
 
     begin
-      response_raw = LinkedData::Client::HTTP.post("#{ADMIN_URL}clear_http_cache", params, raw: true)
-      response[:success] = "HTTP cache successfully flushed"
-    rescue Exception => e
+      LinkedData::Client::HTTP.post("#{ADMIN_URL}clear_http_cache", params, raw: true)
+      response[:success] = 'HTTP cache successfully flushed'
+    rescue StandardError => e
       response[:errors] = "Problem flushing the HTTP cache - #{e.class}: #{e.message}"
     end
-    render :json => response
+    render json: response
   end
 
   def ontologies_report
     response = _ontologies_report
-    render :json => response
+    render json: response
   end
 
   def refresh_ontologies_report
-    response = {errors: '', success: ''}
+    response = { errors: '', success: '' }
 
     begin
       response_raw = LinkedData::Client::HTTP.post(ONTOLOGIES_URL, params, raw: true)
-      response_json = JSON.parse(response_raw, :symbolize_names => true)
+      response_json = JSON.parse(response_raw, symbolize_names: true)
 
       if response_json[:errors]
         _process_errors(response_json[:errors], response, true)
       else
         response = response_json
 
-        if params["ontologies"].nil? || params["ontologies"].empty?
-          response[:success] = "Refresh of ontologies report started successfully";
+        if params['ontologies'].blank?
+          response[:success] = 'Refresh of ontologies report started successfully'
         else
-          ontologies = params["ontologies"].split(",").map {|o| o.strip}
-          response[:success] = "Refresh of report for ontologies: #{ontologies.join(", ")} started successfully";
+          ontologies = params['ontologies'].split(',').map { |o| o.strip }
+          response[:success] = "Refresh of report for ontologies: #{ontologies.join(', ')} started successfully"
         end
       end
-    rescue Exception => e
+    rescue StandardError => e
       response[:errors] = "Problem refreshing report - #{e.class}: #{e.message}"
-      # puts "#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}"
     end
-    render :json => response
+    render json: response
   end
 
   def process_ontologies
@@ -170,20 +172,20 @@ class AdminController < ApplicationController
   end
 
   def delete_submission
-    response = {errors: '', success: ''}
+    response = { errors: '', success: '' }
 
     begin
-      ont = params["acronym"]
+      ont = params['acronym']
       ontology = LinkedData::Client::Models::Ontology.find_by_acronym(ont).first
 
       if ontology
         submissions = ontology.explore.submissions
-        submission = submissions.select {|o| o.submissionId == params["id"].to_i}.first
+        submission = submissions.select { |o| o.submissionId == params['id'].to_i }.first
 
         if submission
           error_response = submission.delete
 
-          if error_response
+          if response_error?(error_response)
             errors = response_errors(error_response) # see application_controller::response_errors
             _process_errors(errors, response, true)
           else
@@ -195,17 +197,16 @@ class AdminController < ApplicationController
       else
         response[:errors] << "Ontology #{ont} was not found in the system"
       end
-    rescue Exception => e
+    rescue StandardError => e
       response[:errors] << "Problem deleting submission #{params["id"]} for ontology #{ont} - #{e.class}: #{e.message}"
     end
-    render :json => response
+    render json: response
   end
 
   def users
     response = _users
-    render :json => response
+    render json: response
   end
-
 
   private
 
@@ -214,12 +215,12 @@ class AdminController < ApplicationController
   end
 
   def _ontologies_report
-    response = {ontologies: Hash.new, report_date_generated: REPORT_NEVER_GENERATED, errors: '', success: ''}
+    response = { ontologies: {}, report_date_generated: REPORT_NEVER_GENERATED, errors: '', success: '' }
     start = Time.now
 
     begin
       ontologies_data = LinkedData::Client::HTTP.get(ONTOLOGIES_URL, {}, raw: true)
-      ontologies_data_parsed = JSON.parse(ontologies_data, :symbolize_names => true)
+      ontologies_data_parsed = JSON.parse(ontologies_data, symbolize_names: true)
 
       if ontologies_data_parsed[:errors]
         _process_errors(ontologies_data_parsed[:errors], response, true)
@@ -228,45 +229,44 @@ class AdminController < ApplicationController
         response[:success] = "Report successfully regenerated on #{ontologies_data_parsed[:report_date_generated]}"
         LOG.add :debug, "Ontologies Report - retrieved #{response[:ontologies].length} ontologies in #{Time.now - start}s"
       end
-    rescue Exception => e
+    rescue StandardError => e
       response[:errors] = "Problem retrieving ontologies report - #{e.message}"
     end
     response
   end
 
-  def _process_errors(errors, response, remove_trailing_comma=true)
+  def _process_errors(errors, response, remove_trailing_comma = true)
     if errors.is_a?(Hash)
       errors.each do |_, v|
-        if v.kind_of?(Array)
-          response[:errors] << v.join(", ")
-          response[:errors] << ", "
+        if v.is_a?(Array)
+          response[:errors] << v.join(', ')
+          response[:errors] << ', '
         else
           response[:errors] << "#{v}, "
         end
       end
-    elsif errors.kind_of?(Array)
-      errors.each {|err| response[:errors] << "#{err}, "}
+    elsif errors.is_a?(Array)
+      errors.each { |err| response[:errors] << "#{err}, " }
     end
     response[:errors] = response[:errors][0...-2] if remove_trailing_comma
   end
 
-  def _delete_ontology(ontology, params)
+  def _delete_ontology(ontology)
     error_response = ontology.delete
     error_response
   end
 
   def _process_ontology(ontology, params)
-    error_response = LinkedData::Client::HTTP.put(ONTOLOGY_URL.call(ontology.acronym), params)
-    error_response
+    LinkedData::Client::HTTP.put(ONTOLOGY_URL.call(ontology.acronym), params)
   end
 
   def _process_ontologies(success_keyword, error_keyword, process_proc)
-    response = {errors: '', success: ''}
+    response = { errors: '', success: '' }
 
-    if params["ontologies"].nil? || params["ontologies"].empty?
+    if params['ontologies'].blank?
       response[:errors] = "No ontologies parameter passed. Syntax: ?ontologies=ONT1,ONT2,...,ONTN"
     else
-      ontologies = params["ontologies"].split(",").map {|o| o.strip}
+      ontologies = params['ontologies'].split(',').map { |o| o.strip }
 
       ontologies.each do |ont|
         begin
@@ -274,8 +274,7 @@ class AdminController < ApplicationController
 
           if ontology
             error_response = self.send(process_proc, ontology, params)
-
-            if error_response
+            if response_error?(error_response)
               errors = response_errors(error_response) # see application_controller::response_errors
               _process_errors(errors, response, false)
             else
@@ -284,28 +283,27 @@ class AdminController < ApplicationController
           else
             response[:errors] << "Ontology #{ont} was not found in the system, "
           end
-        rescue Exception => e
+        rescue StandardError => e
           response[:errors] << "Problem #{error_keyword} ontology #{ont} - #{e.class}: #{e.message}, "
         end
       end
       response[:success] = response[:success][0...-2] unless response[:success].empty?
       response[:errors] = response[:errors][0...-2] unless response[:errors].empty?
     end
-    render :json => response
+    render json: response
   end
 
   def _users
-    response = {users: Hash.new , errors: '', success: ''}
+    response = { users: {}, errors: '', success: '' }
     start = Time.now
     begin
-      response[:users] = JSON.parse(LinkedData::Client::HTTP.get(USERS_URL, {include: 'all'}, raw: true))
+      response[:users] = JSON.parse(LinkedData::Client::HTTP.get(USERS_URL, { include: 'all' }, raw: true))
 
       response[:success] = "users successfully retrieved in  #{Time.now - start}s"
-    LOG.add :debug, "Users - retrieved #{response[:users].length} users in #{Time.now - start}s"
-    rescue Exception => e
+      LOG.add :debug, "Users - retrieved #{response[:users].length} users in #{Time.now - start}s"
+    rescue StandardError => e
       response[:errors] = "Problem retrieving users  - #{e.message}"
     end
     response
   end
-
 end
