@@ -169,14 +169,24 @@ class ApplicationController < ActionController::Base
     file_exists
   end
 
+  def parse_response_body(response)
+    return nil if response.nil?
+
+    if response.respond_to?(:errors) && response.errors
+      response
+    else
+      OpenStruct.new(JSON.parse(response.body, symbolize_names: true))
+    end
+  end
+
   def response_errors(error_struct)
+    error_struct = parse_response_body(error_struct)
     errors = {error: "There was an error, please try again"}
     return errors unless error_struct
     return errors unless error_struct.respond_to?(:errors)
     errors = {}
-    error_struct.errors.each {|e| ""}
     error_struct.errors.each do |error|
-      if error.is_a?(Struct)
+      if error.is_a?(OpenStruct) || error.is_a?(Struct)
         errors.merge!(struct_to_hash(error))
       else
         errors[:error] = error
@@ -185,11 +195,25 @@ class ApplicationController < ActionController::Base
     errors
   end
 
+  def response_success?(response)
+    return true if response.nil?
+
+    if response.respond_to?(:status) && response.status
+        response.status.to_i < 400
+    else
+      !(response.respond_to?(:errors) && response.errors)
+    end
+  end
+
+  def response_error?(response)
+    !response_success?(response)
+  end
+
   def struct_to_hash(struct)
     hash = {}
     struct.members.each do |attr|
       next if [:links, :context].include?(attr)
-      if struct[attr].is_a?(Struct)
+      if struct[attr].is_a?(Struct) || struct[attr].is_a?(OpenStruct)
         hash[attr] = struct_to_hash(struct[attr])
       else
         hash[attr] = struct[attr]
@@ -204,16 +228,6 @@ class ApplicationController < ActionController::Base
 
   def redirect_to_home # Redirect to Home Page
     redirect_to "/"
-  end
-
-  def redirect_to_history # Redirects to the correct tab through the history system
-    if session[:redirect].nil?
-      redirect_to_home
-    else
-      tab = find_tab(session[:redirect][:ontology])
-      session[:redirect]=nil
-      redirect_to uri_url(:ontology=>tab.ontology_id,:conceptid=>tab.concept)
-    end
   end
 
   def redirect_new_api(class_view = false)
@@ -299,40 +313,20 @@ class ApplicationController < ActionController::Base
     session[:user] && session[:user].admin?
   end
 
-  # updates the 'history' tab with the current selected concept
   def update_tab(ontology, concept)
-    array = session[:ontologies] || []
+    onts = session[:ontologies] || []
     found = false
-    for item in array
-      if item.ontology_id.eql?(ontology.id)
-        item.concept=concept
-        found=true
+    onts.each do |ont|
+      if ont.ontology_id.eql? ontology.id
+        ont.concept = concept
+        found = true
       end
     end
 
-    unless found
-      array << History.new(ontology.id, ontology.name, ontology.acronym, concept)
-    end
+    onts << History.new(ontology.id, ontology.name, ontology.acronym, concept) unless found
 
-    session[:ontologies]=array
-  end
-
-  # Removes a 'history' tab
-  def remove_tab(ontology_id)
-    array = session[:ontologies]
-    array.delete(find_tab(ontology_id))
-    session[:ontologies]=array
-  end
-
-  # Returns a specific 'history' tab
-  def find_tab(ontology_id)
-    array = session[:ontologies]
-    for item in array
-      if item.ontology_id.eql?(ontology_id)
-        return item
-      end
-    end
-    return nil
+    # The "Recently Viewed" menu item displays the contents of session[:ontologies]
+    session[:ontologies] = onts
   end
 
   def check_delete_mapping_permission(mappings)
