@@ -1,28 +1,28 @@
+# frozen_string_literal: true
+
 # VirtualApplianceIdValidator
 #
-# Custom validator for virtual appliance IDs. The virtual appliance ID in a 
-# license key must match the ID of the virtual appliance against which a 
+# Custom validator for virtual appliance IDs. The virtual appliance ID in a
+# license key must match the ID of the virtual appliance against which a
 # license is submitted.
 #
 #   class License < ApplicationRecord
 #     validates :encrypted_key, virtual_appliance_id: true
 #   end
 #
-
 class VirtualApplianceIdValidator < ActiveModel::EachValidator
-
   def validate_each(record, attribute, value)
     return if record.is_trial?
 
     response = JSON.parse(
-      LinkedData::Client::HTTP.get(LinkedData::Client.settings.rest_url + "/admin/update_info", {}, raw: true)
+      LinkedData::Client::HTTP.get("#{LinkedData::Client.settings.rest_url}/admin/update_info", {}, raw: true)
     )
-    if response["error"]
+    if response['appliance_id'].blank?
       record.errors.add(attribute, :no_appliance_id_for_comparison)
       return false
     end
-    
-    unless valid_virtual_appliance_id?(value, response["appliance_id"])
+
+    unless valid_virtual_appliance_id?(value, response['appliance_id'])
       record.errors.add(attribute, :appliance_id_mismatch)
     end
   end
@@ -36,5 +36,4 @@ class VirtualApplianceIdValidator < ActiveModel::EachValidator
   rescue OpenSSL::Cipher::CipherError, OpenSSL::PKey::RSAError
     false
   end
-
 end
